+++ lib/src/components/export.dart
@@ -9,6 +9,7 @@ export 'country_picker_pro2/country_picker_pro.dart';
 
 // IMAGES
 export 'images/nbx_network_image.dart';
+export 'images/configs/export.dart';
 
 // INPUTS
 export 'inputs/export.dart';