+++ lib/src/components/inputs/text_fields/nbx_text_field_with_state_widget.dart
@@ -31,7 +31,6 @@ class _NbxTextFieldWithStateWidgetState
     super.dispose();
     if (_parameters.autoDisposeController) {
       _parameters.controller?.dispose();
-      debugPrint('[${_parameters.labelText}] - Has be deleted');
     }
   }